+++ Makerspace/Makerspace/Content/bootstrap.css
@@ -384,7 +384,22 @@ h5, .h5 {
 h6, .h6 {
     font-size: 1rem;
 }
+.fs-1 {
+    font-size: 2.5rem !important;
+}
 
+.fs-2 {
+    font-size: 2rem !important;
+}
+.fs-3 {
+    font-size: 1.75rem !important;
+}
+.fs-4 {
+    font-size: 1.5rem !important;
+}
+.fs-5 {
+    font-size: 1.25rem !important;
+}
 .lead {
     font-size: 1.25rem;
     font-weight: 300;
