+++ Makerspace/Makerspace/EquipmentDetail.aspx
@@ -2,14 +2,12 @@
 <asp:Content ID="EDetail" ContentPlaceHolderID="MainContent" runat="server">
     <div class="wrapper">
         <div class="one">
-            <div>          
-                <img src="Images\A18-030.jpg" width="260" height="260" style="vertical-align:baseline">
-            <br />
-            <div style="display: flex; justify-content:center">
-                 <asp:Label id="name" style="background-color: #00196E; color: white; font-size: 24px; border-radius: 4px; font-weight: 700; position: absolute; padding: 6px 15px 6px 15px;" runat="server">Heavy Machine</asp:Label>
+            <div >          
+                <img src="Images\A18-030.jpg" width="260" height="260" style="vertical-align:baseline" class="rounded">
             </div>
-            </div>
-  
+            <div style="margin-top:1.625rem;">
+                 <asp:Label id="name" style="background-color: #00196E; color: white; font-size: 24px; border-radius: 4px; font-weight: 700; padding: 6px 15px 6px 15px;" runat="server">Heavy Machine</asp:Label>
+            </div>  
         </div>
         <div class="two">
             <asp:Label id="eid" style="background-color:#FFAD1D; color:#00196E; padding:10px 30px 10px 30px; font-size:24px; font-weight:700; text-align:center; border-radius:4px; vertical-align: super" runat="server" OnLoad="code_Load"></asp:Label>
@@ -28,4 +26,47 @@
             <br />
         </div>
     </div>
+    <div style="margin-bottom: 6.25rem"></div>
+    <%-- startregion: Subitem Listview --%>
+    <asp:ListView runat="server"  ID="Subitem_Listview" GroupItemCount="2">
+        <GroupTemplate>
+            <tr id="itemPlaceholderContainer" runat="server" style="min-width:1140px; display: flex;justify-content: space-between">
+                <td id="itemPlaceholder" runat="server" ></td>
+            </tr>
+        </GroupTemplate>
+        <ItemTemplate>
+            <td runat="server">
+                <table class="rounded" style="background-color: #F1F3F6; min-width:400px;">
+                    <tr>
+                        <td rowspan="2" style="width:104px" class="p-3">
+                            <div style="background-color:#F1F3F6;"><img src="Images/<%#Eval("category_src") %>" height="82" width="82"></div>
+                        </td>
+                        <td class="py-0 pl-0 pr-3 align-bottom">
+                            <asp:Label runat="server" Text='<%# String.Concat(Eval("name"), " #", Eval("num"))%>' CssClass="fs-5 text-blue-fuv font-weight-medium"></asp:Label>  
+                        </td>
+                    </tr>
+                    <tr>
+                        <td class="py-0 pl-0 pr-3 align-top">
+                            <asp:Label runat="server" Text='<%# String.Concat("Location: ", Eval("room_space_name"), " - ", Eval("object_name"), " ", Eval("object_num"))%>' ForeColor="#3A4276"></asp:Label>
+                        </td>
+                    </tr>          
+                </table>
+            </td>
+        </ItemTemplate>
+        <LayoutTemplate>
+            <table  style="width:100%;">
+                <tbody>
+                    <tr>
+                        <td>
+                            <table class="table table-borderless" id="groupPlaceholderContainer" runat="server" style="width:100%">
+                                <tr id="groupPlaceholder"></tr>
+                            </table>
+                        </td>
+                    </tr>
+                </tbody>
+            </table>               
+        </LayoutTemplate>         
+    </asp:ListView>
+    <%-- endregion: Subitem Listview --%>
+
 </asp:Content>
