+++ Makerspace/Makerspace/Makerspace.csproj
@@ -169,7 +169,6 @@
     <Content Include="Images\logo.png" />
     <Content Include="Profile.aspx" />
     <Content Include="Project.aspx" />
-    <Content Include="ResourceDetail.aspx" />
     <Content Include="Scripts\bootstrap.bundle.js" />
     <Content Include="Scripts\bootstrap.bundle.min.js" />
     <Content Include="Scripts\bootstrap.js" />
@@ -366,12 +365,6 @@
       <DependentUpon>Project.aspx</DependentUpon>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ResourceDetail.aspx.designer.cs">
-      <DependentUpon>ResourceDetail.aspx</DependentUpon>
-    </Compile>
-    <Compile Include="ResourceDetail.aspx.cs">
-      <SubType>ASPXCodeBehind</SubType>
-    </Compile>
     <Compile Include="Site.Master.cs">
       <DependentUpon>Site.Master</DependentUpon>
       <SubType>ASPXCodeBehind</SubType>
