--- Makerspace/Makerspace/ResourceDetail.aspx.cs
@@ -1,48 +1,0 @@
-﻿using System;
-using System.Configuration;
-using System.Data;
-using System.Data.SqlClient;
-using System.Web.UI;
-using System.Web.UI.WebControls;
-
-
-namespace Makerspace
-{
-    public partial class ResourceDetail : System.Web.UI.Page
-    {
-        protected static int ItemsCount = 0;
-        protected static string CONSTRING = ConfigurationManager.ConnectionStrings["MakerspaceDBConnectionString"].ConnectionString;
-        protected void Page_Load(object sender, EventArgs e)
-        {
-            if (!IsPostBack)
-            {
-                load();
-            }
-            else
-            {
-
-            }
-
-        }
-
-        //Post back data
-        //Bind ListView
-        protected void load()
-        {
-            using (SqlConnection con = new SqlConnection(CONSTRING))
-            {
-                SqlCommand cmd = new SqlCommand("uspReadAllEquipment");
-                cmd.CommandType = CommandType.StoredProcedure;
-                con.Open();
-                cmd.Connection = con;
-                DataTable db = new DataTable();
-                SqlDataAdapter adapter = new SqlDataAdapter(cmd);
-                adapter.Fill(db);
-                ResourceDetail_Listview.DataSource = db;
-                ResourceDetail_Listview.DataBind();
-            }
-        }
-
-    }
-
-}
