+++ Makerspace/Makerspace/Site.Master
@@ -54,9 +54,8 @@
             grid-row-start:1;
             grid-row-end: 3;
             display:flex;
-            align-self:flex-start;
-            align-items: flex-end;
-            justify-content: center;
+            flex-direction:column;
+            align-items:center;
         }
 
         .two {
@@ -67,7 +66,7 @@
         .three {
            grid-column-start:2;
            grid-row-start: 2;
-           height: 300px;
+           /*height: 300px;*/
         }
 
         .equipment_info {
@@ -83,7 +82,7 @@
     <br />
     <div class="auto-style2" role="navigation" style="background-color:#00196E">
         <div class="auto-style1">
-            <a class="navbar-brand" style="color:white; font-weight:800" runat="server" href="~/Default.aspx">Makerspace</a>
+            <a class="navbar-brand" style="color:white; font-weight:800" runat="server" href="~/Resource.aspx">Makerspace</a>
             <div class="navbar-nav mr-auto" style="width:100%; list-style-type:none; padding:0px; display:inline; flex-direction:row;text-align:justify">
                 <%--<a runat="server" class="navbar-items" href="~/">Home</a> --%>
                 <a runat="server" class="navbar-items" href="~/About">About</a>
